% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.precision.R
\name{plot.precision}
\alias{plot.precision}
\title{Plot Bias vs Variance for Ridge Precision}
\usage{
\method{plot}{precision}(
  x,
  xvar = "norm.beta",
  yvar = c("det", "trace", "max.eig"),
  labels = c("lambda", "df"),
  criteria = NULL,
  pch = 16,
  cex = 1.5,
  col,
  main = NULL,
  xlab,
  ylab,
  ...
)
}
\arguments{
\item{x}{A data frame of class \code{"precision"} resulting from \code{link{precision}} called
on a \code{"ridge"} object. Named \code{x} only to conform with the \code{\link{plot}} generic.}

\item{xvar}{The character name of the column to be used for the horizontal axis.}

\item{yvar}{The character name of the column to be used for the vertical axis.}

\item{labels}{The character name of the column to be used for point labels.}

\item{criteria}{The vector of optimal shrinkage criteria from the \code{link{ridge}} call to be added
as labels in the plot. Not yet implemented.}

\item{pch}{Plotting character for points}

\item{cex}{Character size for points}

\item{col}{Point colors}

\item{main}{Plot title}

\item{xlab}{Label for horizontal axis}

\item{ylab}{Label for vertical axis}

\item{...}{Other arguments passed to \code{link{plot}}.}
}
\value{
Returns nothing. Used for the side effect of plotting.
}
\description{
Plot Bias vs Variance for Ridge Precision
}
\examples{
# none yet
}
